Updated dir_imgcache.js
Dramatically increased the speed of fetching an image compared to my last version. The first time an image is got it will take the same time as previously (Which in my test could range anywhere between 200 - 900ms due to diving into the FileSystem), however I then store a reference to that image in a sessionStorage value (This could be changed to localStorage if you want to persist between app loads) which I then attempt to get from with each image request. This takes the time down to around 2ms for every additional request for that image.